backend/requirements.txt
# Backend Requirements
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
openai==1.3.7
pydantic==2.5.0
httpx==0.25.2

# Development
pytest==7.4.3